# admin-server/swagger.yaml
openapi: 3.0.0
info:
  title: Admin API
  description: API для управления товарами
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /products:
    get:
      summary: Получить все товары
      responses:
        '200':
          description: Список товаров
    post:
      summary: Добавить новый товар
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Товар добавлен

  /products/{id}:
    get:
      summary: Получить товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о товаре
        '404':
          description: Товар не найден

    put:
      summary: Обновить товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Товар обновлен
        '404':
          description: Товар не найден

    delete:
      summary: Удалить товар по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Товар удален
        '404':
          description: Товар не найден